requirements.txt
Django==3.1
django-crispy-forms==1.9.2
six==1.15.0
Pillow
django-widget-tweaks==1.4.8
django-decouple==2.1
unipath==1.1
dj-database-url==0.5.0
django-countries==6.1.3
django-multiselectfield==0.1.12
django-formtools==2.2
django-tinymce==3.1.0
django-import-export==2.4.0